Return empty dict in controller(Urgent)
1. What is the problem?
After updated to version 1.2, the WSGI framework we use, pecan has a
behaviour that if the handle function return None, the setting of
resonse status code will not take effect. In the deletion handle
functions of pod and pod-binding controllers, we set the response
status code but return None, thus the status code is not changed and
our tests fail.

2. What is the solution to the problem?
Return an empty dict in the deletion handle functions.

3. What the features need to be implemented to the Tricircle to realize
the solution?
No new features.

Change-Id: Idc24b022290e095ea2435b48b8d34b678611777e