Implement force detach volume
1.what is the problem?
In Trio2o, when a volume is in a stuck 'attaching' or 'detaching', Pod
can not detaching it by nova-apigw 'detach volume', so that this volume can
not reuse by ohther Pod. In addition, it may cause un-synchronization between
the Cinder database and the backend storage.
   There is an user case in nova[1]. It may also occur on a trio2o Pod.
We don't need to worry about its security, because the Cinder api
'os-force-detach' is for safe cleanup. More user case and information
about 'force detach' in this document[2].

2.what is the solution to the problem?
We use Cinder api "os-force-detach" to detach it.

3. What the features need to be implemented to the Trio2o
   to realize the solution?
Implement the force detach server action.

[1] https://blueprints.launchpad.net/nova/+spec/add-force-detach-to-nova
[2] https://github.com/openstack/cinder-specs/blob/master/specs/liberty/implement-force-detach-for-safe-cleanup.rst

Change-Id: Iea15c4d9ce469fb640e7e98c1859a18fe12ce7f3