Fixed trio2o tempest plugin config
The trio2o tempest plugin is using the trio2o project config
which overrides the tempest config while running tempest plugin tests
with tempest. It leads to the failure of tempest command. We need to
use oslo_config for specifiying tempest plugin config for trio2o
to avoid overriding and avoid the above failure.

Change-Id: Ia8a165d496c6520313f1454d11f98adb8830d027
Closes-Bug: #1692291